Move CLI to packages
