+++ packages/cli/commands/deps_commands/get.js
@@ -1,4 +1,4 @@
-const {fetchDependencies} = require("../../../package/index");
+const { fetchDependencies } = require("../../../../package/index");
 
 exports.command = "get";
 exports.desc = "Download every dependency listed in the package config file";
