+++ packages/cli/commands/highlight.js
