+++ packages/cli/commands/tests/highlight.test.js
