+++ packages/cli/commands/tests/new.test.js
@@ -2,8 +2,11 @@ const tmp = require("tmp");
 const fs = require("fs");
 const path = require("path");
 
-const { _new } = require("../");
-const { CONFIGFILE_NAME, getPackageConfig } = require("../../../package/index");
+const { _new } = require("..");
+const {
+  CONFIGFILE_NAME,
+  getPackageConfig
+} = require("../../../../package/index");
 
 test("Create a package", async () => {
   const dir = tmp.dirSync();
