+++ packages/cli/lib/colors.js
