+++ packages/cli/lib/platforms.js
