+++ packages/cli/lib/progress.js
