+++ packages/cli/lib/tests/colors.test.js
