+++ packages/cli/lib/tests/progress.test.js
