+++ packages/cli/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "cli",
+  "version": "0.1.0",
+  "description": "",
+  "main": "index.js",
+  "directories": {
+    "lib": "lib"
+  },
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "Garrit Franke <garritfranke@gmail.com>",
+  "license": "Apache-2.0",
+  "dependencies": {
+    "clio-core": "^0.1.0"
+  }
+}
